Move SSH tests into a single-threaded process
Necessary because LibSSH is not thread-safe.
